[refactor] RandomNumberUtil 관련 인터페이스 및 구현체 구조 변경
