+++ src/main/java/racingcar/Application.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import racingcar.controller.GameController;
 import racingcar.domain.MovingStrategy;
-import racingcar.util.RandomNumberGenerator;
+import racingcar.util.RandomNumberUtilImp;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -10,7 +10,7 @@ public class Application {
     public static void main(String[] args) {
         GameController gameController = new GameController(new InputView(),
                 new OutputView(),
-                new MovingStrategy(new RandomNumberGenerator()));
+                new MovingStrategy(new RandomNumberUtilImp()));
         gameController.play();
     }
 }
