+++ src/main/java/racingcar/domain/MovingStrategy.java
@@ -1,5 +1,7 @@
 package racingcar.domain;
 
+import racingcar.util.RandomNumberUtil;
+
 import static racingcar.domain.MovingStrategySetting.*;
 
 public class MovingStrategy {
