+++ src/main/java/racingcar/util/RandomNumberUtil.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.util;
 
 public interface RandomNumberUtil {
     int generate(int startRangeNumber, int endRangeNumber);
