+++ src/test/java/racingcar/domain/GameRoundTest.java
@@ -2,7 +2,6 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.util.RandomNumberGenerator;
 
 import static org.junit.jupiter.api.Assertions.*;
