+++ src/test/java/racingcar/domain/MovingStrategyTest.java
@@ -1,9 +1,7 @@
 package racingcar.domain;
 
-import camp.nextstep.edu.missionutils.Randoms;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.util.RandomNumberGenerator;
 
 import static org.junit.jupiter.api.Assertions.*;
